rubato
Module asynchro_fast

Macros

t 🔒

Structs

FastFixedIn

An asynchronous resampler that accepts a fixed number of audio frames for input and returns a variable number of frames.

FastFixedOut

An asynchronous resampler that returns a fixed number of audio frames. The number of input frames required is given by the input_frames_next function.

Enums

PolynomialDegree

Degree of the polynomial used for interpolation. A higher degree gives a higher quality result, while taking longer to compute.

Constants

POLYNOMIAL_LEN_I 🔒

POLYNOMIAL_LEN_U 🔒

Functions

interp_cubic 🔒

Perform cubic polynomial interpolation to get value at x. Input points are assumed to be at x = -1, 0, 1, 2.

interp_lin 🔒

Linear interpolation between two points at x=0 and x=1.

interp_quintic 🔒

Perform quintic polynomial interpolation to get value at x. Input points are assumed to be at x = -2, -1, 0, 1, 2, 3.

interp_septic 🔒

Perform septic polynomial interpolation to get value at x. Input points are assumed to be at x = -3, -2, -1, 0, 1, 2, 3, 4.

validate_ratios 🔒