rubato
Module asynchro_sinc

Structs

SincFixedIn

An asynchronous resampler that accepts a fixed number of audio frames for input and returns a variable number of frames.

SincFixedOut

An asynchronous resampler that returns a fixed number of audio frames. The number of input frames required is given by the input_frames_next function.

SincInterpolationParameters

A struct holding the parameters for sinc interpolation.

Enums

SincInterpolationType

Interpolation methods that can be selected. For asynchronous interpolation where the ratio between input and output sample rates can be any number, it’s not possible to pre-calculate all the needed interpolation filters. Instead they have to be computed as needed, which becomes impractical since the sincs are very expensive to generate in terms of cpu time. It’s more efficient to combine the sinc filters with some other interpolation technique. Then, sinc filters are used to provide a fixed number of interpolated points between input samples, and then, the new value is calculated by interpolation between those points.

Functions

interp_cubic 🔒

Perform cubic polynomial interpolation to get value at x. Input points are assumed to be at x = -1, 0, 1, 2.

interp_lin 🔒

Perform linear interpolation between two points at x=0 and x=1.

interp_quad 🔒

Perform quadratic polynomial interpolation to get value at x. Input points are assumed to be at x = 0, 1, 2.

make_interpolator

validate_ratios 🔒