rubato::asynchro_sinc
Struct SincFixedOut

pub struct SincFixedOut<T> {

    nbr_channels: usize,
    chunk_size: usize,
    needed_input_size: usize,
    last_index: f64,
    current_buffer_fill: usize,
    resample_ratio: f64,
    resample_ratio_original: f64,
    target_ratio: f64,
    max_relative_ratio: f64,
    interpolator: Box<dyn SincInterpolator<T>>,
    buffer: Vec<Vec<T>>,
    interpolation: SincInterpolationType,
    channel_mask: Vec<bool>,

}

An asynchronous resampler that returns a fixed number of audio frames. The number of input frames required is given by the input_frames_next function.

The resampling is done by creating a number of intermediate points (defined by oversampling_factor) by sinc interpolation. The new samples are then calculated by interpolating between these points.

The resampling ratio can be freely adjusted within the range specified to the constructor. Adjusting the ratio does not recalculate the sinc functions used by the anti-aliasing filter. This causes no issue when increasing the ratio (which slows down the output). However when decreasing more than a few percent (i.e. speeding up the output), the filters can no longer suppress all aliasing and this may lead to some artefacts. Higher maximum ratios require more memory to be allocated by input_buffer_allocate and an internal buffer.

Fields
nbr_channels: usizechunk_size: usizeneeded_input_size: usizelast_index: f64current_buffer_fill: usizeresample_ratio: f64resample_ratio_original: f64target_ratio: f64max_relative_ratio: f64interpolator: Box<dyn SincInterpolator<T>>buffer: Vec<Vec<T>>interpolation: SincInterpolationTypechannel_mask: Vec<bool>
Implementations

impl<T> SincFixedOut<T>
where T: Sample,

pub fn new( resample_ratio: f64, max_resample_ratio_relative: f64, parameters: SincInterpolationParameters, chunk_size: usize, nbr_channels: usize, ) -> Result<Self, ResamplerConstructionError>

Create a new SincFixedOut.

Parameters are:

resample_ratio: Starting ratio between output and input sample rates, must be > 0.

max_resample_ratio_relative: Maximum ratio that can be set with Resampler::set_resample_ratio relative to resample_ratio, must be >= 1.0. The minimum relative ratio is the reciprocal of the maximum. For example, with max_resample_ratio_relative of 10.0, the ratio can be set between resample_ratio * 10.0 and resample_ratio / 10.0.

parameters: Parameters for interpolation, see SincInterpolationParameters.

chunk_size: Size of output data in frames.

nbr_channels: Number of channels in input/output.

pub fn new_with_interpolator( resample_ratio: f64, max_resample_ratio_relative: f64, interpolation_type: SincInterpolationType, interpolator: Box<dyn SincInterpolator<T>>, chunk_size: usize, nbr_channels: usize, ) -> Result<Self, ResamplerConstructionError>

Create a new SincFixedOut using an existing Interpolator.

Parameters are:

resample_ratio: Starting ratio between output and input sample rates, must be > 0.

max_resample_ratio_relative: Maximum ratio that can be set with Resampler::set_resample_ratio relative to resample_ratio, must be >= 1.0. The minimum relative ratio is the reciprocal of the maximum. For example, with max_resample_ratio_relative of 10.0, the ratio can be set between resample_ratio * 10.0 and resample_ratio / 10.0.

interpolation_type: Parameters for interpolation, see SincInterpolationParameters.

interpolator: The interpolator to use.

chunk_size: Size of output data in frames.

nbr_channels: Number of channels in input/output.