rubato::sample
Trait CoerceFrom

pub trait CoerceFrom<T> {
    // Required method
    fn coerce_from(value: T) -> Self;
}

The trait used to coerce a value infallibly from one type to another.

This is similar to doing value as T where T is a floating point type. Loss of precision may happen during coercions if the coerced from value doesn’t fit fully within the target type.

Required Methods

fn coerce_from(value: T) -> Self

Perform a coercion from value into the current type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.