rubato::sinc_interpolator
Trait SincInterpolator

pub trait SincInterpolator<T>: Send {
    // Required methods
    fn get_sinc_interpolated(
        &self,
        wave: &[T],
        index: usize,
        subindex: usize,
    ) -> T;
    fn len(&self) -> usize;
    fn nbr_sincs(&self) -> usize;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

Functions for making the scalar product with a sinc.

Required Methods

fn get_sinc_interpolated(&self, wave: &[T], index: usize, subindex: usize) -> T

Make the scalar product between the waveform starting at index and the sinc of subindex.

fn len(&self) -> usize

Get sinc length.

fn nbr_sincs(&self) -> usize

Get number of sincs used for oversampling.

Provided Methods

fn is_empty(&self) -> bool

Check if sincs are empty.