rubato
Struct FftFixedInOut

pub struct FftFixedInOut<T> {
    nbr_channels: usize,
    chunk_size_in: usize,
    chunk_size_out: usize,
    fft_size_in: usize,
    channel_mask: Vec<bool>,
    overlaps: Vec<Vec<T>>,
    resampler: FftResampler<T>,
}

A synchronous resampler that accepts a fixed number of audio frames for input and returns a fixed number of frames.

The resampling is done by FFT:ing the input data. The spectrum is then extended or truncated as well as multiplied with an antialiasing filter before it’s inverse transformed to get the resampled waveforms.

Fields
nbr_channels: usizechunk_size_in: usizechunk_size_out: usizefft_size_in: usizechannel_mask: Vec<bool>overlaps: Vec<Vec<T>>resampler: FftResampler<T>
Implementations

impl<T> FftFixedInOut<T>
where T: Sample,

pub fn new( sample_rate_input: usize, sample_rate_output: usize, chunk_size_in: usize, nbr_channels: usize, ) -> Result<Self, ResamplerConstructionError>

Create a new FftFixedInOut.

Parameters are:

sample_rate_input: Input sample rate, must be > 0.

sample_rate_output: Output sample rate, must be > 0.

chunk_size_in: desired length of input data in frames, actual value may be different.

nbr_channels: number of channels in input/output.